# By xeq (10) and 应雨洋 (8)
# Via 应雨洋 (6) and xeq (4)
* 'develop-bug-fix2' of https://github.com/LoyoChengdu/loyo-android:
  销售机会权限调整
  修复 bug
  订单详情 title 功能按钮权限  订单状态的图片颜色
  赢单机会 生成 订单 一次完成
  1.修复动态字段传值问题 2.修复新建编辑订单，购买产品第二次没回显bug
  客户 的订单 模块 一次完成
  1.新建，编辑动态字段必填判断 2.订单编辑完成
  回款审批 基本完成
  订单审批相关 调整
  回款审批一次 完成
  代码合并
  订单审批 一次完成
  编辑回款记录，接口调通
  回款记录删除接口调通
  创建回款记录接口，获取回款记录列表接口，调完，数据绑定
  订单详情的删除 编辑 意外终止
  1.删除回款计划接口，编辑回款计划接口，调通 2.回款计划整体流程完成
  代码合并

+++ app/src/main/java/com/loyo/oa/v2/activityui/order/bean/OrderAdd.java
@@ -16,7 +16,7 @@ public class OrderAdd implements Serializable{
     public String customerName;  //对应客户
     public String attachmentUUId;
 
-    public int dealMoney;     //成交金额
+    public float  dealMoney;     //成交金额
     public String orderNum;   //订单编号
     public String remark;     //备注
